Move utils tests into different separate modules
